% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossrunem.R
\name{crossrunem}
\alias{crossrunem}
\title{Joint Probabilities for Crossings and Longest Run,
based on the empirical median}
\usage{
crossrunem(nmax = 100, prec = 120, printn = FALSE)
}
\arguments{
\item{nmax}{max sequence length.}

\item{prec}{mpft precision.}

\item{printn}{logical for including progress output.}
}
\value{
nfi, nfn, list of number of subsets of size m, 0<=m<=n
with first element included (nfi) and not included (nfn).
}
\description{
Joint probability distribution for the number of crossings
C and the longest run L in a sequence of n independent continuous observations
classified as above or below the empirical median. To enhance precision,
results are stored in mpfr arrays and the probabilities are multiplied by
\eqn{choose(n,m)/2} where m=n/2, even n assumed. The probabilities are
integers in this representation.
}
\examples{
em10 <- crossrunem(nmax=10,printn=TRUE)
}
