% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{joint60em}
\alias{joint60em}
\title{Joint probabilities, n=14, based on empirical median}
\format{matrix, 60 rows and 30 columns}
\source{
generated by the function crossrun and
transformed from an Rmpfr array to a matrix
}
\usage{
joint60em
}
\description{
Joint probabilities of the number C of crossings
(1, ... 59) and the longest run L (1, ..., 30) in a
sequence of n=60 independent continuous observations
classified as above or below the empirical median.
The probabilities are stored in the "times" representations, 
multiplied by \eqn{choose(60,30)} and are integers in 
this representation. Only the joint distributions 
for n=14 and 60 are included in the package to avoid excessive 
storage, but all joint distributions for even n up to 64 have
been generated by the function crossrunem.
}
\keyword{datasets}
