% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.r
\name{simclem}
\alias{simclem}
\title{Simulation of observations in n=2m independent and
identical continuous variables classified as above or
below the median.}
\usage{
simclem(m1 = 14, nsim = 1e+05)
}
\arguments{
\item{m1, }{half the number of observations}

\item{nsim}{number of simulations}
}
\value{
a data frame with the number of crossings cs and
longest run ls in each simulation.
}
\description{
Simulation of observations in n=2m independent and
identical continuous variables classified as above or
below the median. The observations are drawn from a standard
normal distruibution, but the results do not depend on the
distribution as long as it is continuous.
}
\examples{
simcl14 <- simclem(m1=14, nsim=100)
mean(simcl14$cs) # mean number of crossings
mean(simcl14$ls) # mean longest run
}
